+++ src/briefy/reflex/config.py
@@ -7,3 +7,8 @@ TASKS_RESULT_DB = config('TASKS_RESULT_DB', default='redis://localhost:6379/2')
 
 # leica
 LEICA_BASE = config('LEICA_BASE', default='http://briefy-leica.briefy-leica')
+
+# alexandria
+ALEXANDRIA_BASE = config('ALEXANDRIA_BASE', default='http://briefy-alexandria.briefy-alexandria')
+ALEXANDRIA_LEICA_ROOT = '48f47fdc-922b-4aae-8388-0fb23a123fcc'
+ALEXANDRIA_CITYPACKAGES_ROOT = '27f0cdff-14da-42c4-880a-51c3d6f0b841'