+++ HISTORY.rst
@@ -6,5 +6,6 @@ History
 ------------------
 
     * Celery code to define app, config, worker and tasks module (rudaporto).
-    * Tasks module to query orders accepted on Leica (rudaporto).
-    * Tasks module to query in google drive folder contents of delivery folder (rudaporto).
+    * Task module to query orders accepted on Leica (rudaporto).
+    * Task module to query in google drive folder contents of delivery folder (rudaporto).
+    * Task module to add collections and assets in briefy.alexandria from Leica Orders payload (rudaporto).